# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bloom & Root Flower Nursery API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  category?: string;
  /** @minimum 0 */
  stock?: number;
  imageUrl?: string;
  featured?: boolean;
  careLevel?: string;
  sunlight?: string;
  watering?: string;
}